Workload bugs fixed

+++ src/main/java/edu/bilkent/cs319/team9/ta_management_system/dto/DutyLogDto.java
@@ -10,7 +10,7 @@ import java.util.Set;
 public class DutyLogDto {
     private Long id;
     private String taskType;
-    private int workload;
+    private float workload;
     private LocalDateTime startTime;
     private LocalDateTime dateTime;
     private LocalDateTime endTime;
